Updated Base Data Acquisition (SMA, see details)
Due to a temporary oversight by the BLM GPB Support team, there was an issue where the all-in-one feature class "SurfaceManagementAgency" was missing from the download, causing the tool to fail. I implemented a fix here to manually generate this feature class from the individual agency-specific datasets IF the all-in-one is missing.